refactor: restructured backend folder for professionalism
